feat: IDP-1106 findById optimize to be run on cosmosDB (#152)
* IDP-1106 findById optimize to be run on cosmosDB

* extending coverage
